Extend module end date in LumiNUS responses
I would've preferred to mock the time during unit tests, but it is impossible to do so properly in
Elixir/Erlang (libraries could call functions themselves that get the time, so then we end up
with different parts of the whole executable seeing different times).

So let's just edit the responses instead.